Add VGG16 execution with cifar10 dataset on the zcu102
We get a pre-trained VGG16 model created by direct training. After that, we pass a cifar10 image as an input to the model to derive execution results on the CPU, VTA simulator, and finally on the zcu102 board. Specifically, the CPU is first used on the host PC to check whether the model is operating normally. Second, we use VTA simulation on the host PC to obtain execution codes through the model-compiler. Finally, we skip the model-compiler on the zcu102 board and deliver the files acquired in the second process through AWS so they can be executed.

Signed-off-by: taehokim20 <taeho.kim@colorado.edu>